scraper/requirements.txt
# Selenium Scraper Dependencies
selenium==4.18.1
webdriver-manager==4.0.1
pandas==2.2.1
python-dotenv==1.0.1
tqdm==4.66.2
requests==2.31.0
